+++ CHANGES.md
@@ -1,6 +1,7 @@
 #### [unreleased]
 * move all Pro CSS to `CSS\Pro`
 * fix CSS for all day multiday week view
+* update hook name for Events Calendar settings to load admin resources
 
 #### 7.0.5 / 2022-05-17
 * load Superpowers JS on mobile too